backend/requirements.txt
fastapi
uvicorn
tqdm
pytesseract
pdf2image
python-multipart
Pillow
numpy
PyMuPDF
jinja2
